+++ sdk/src/test/scala/io/horizen/fixtures/SidechainRelatedMainchainOutputFixture.scala
@@ -1,7 +1,7 @@
 package io.horizen.fixtures
 
 import io.horizen.block.{MainchainTxForwardTransferCrosschainOutput, MainchainTxSidechainCreationCrosschainOutput}
-import io.horizen.proposition.PublicKey25519Proposition
+import io.horizen.proposition.{Proposition, PublicKey25519Proposition}
 import io.horizen.transaction.mainchain.{ForwardTransfer, SidechainCreation}
 import io.horizen.utils.BytesUtils
 import org.scalatestplus.mockito.MockitoSugar.mock
@@ -10,10 +10,12 @@ import scala.util.Random
 
 trait SidechainRelatedMainchainOutputFixture extends SecretFixture {
 
-  def getForwardTransfer(proposition: PublicKey25519Proposition, sidechainId: Array[Byte], seed: Long = 12345): ForwardTransfer = {
+  def getForwardTransfer(proposition: Proposition, sidechainId: Array[Byte], seed: Long = 12345): ForwardTransfer = {
     Random.setSeed(seed)
+    val trailingZeroBytes = new Array[Byte](32 - proposition.bytes().length)
+    val propositionBytes: Array[Byte] = BytesUtils.reverseBytes(proposition.bytes()) ++: trailingZeroBytes
     val output = new MainchainTxForwardTransferCrosschainOutput(new Array[Byte](1), sidechainId,
-      Random.nextInt(10000), BytesUtils.reverseBytes(proposition.bytes()), getMcReturnAddress)
+      Random.nextInt(10000), propositionBytes, getMcReturnAddress)
     val forwardTransferHash = new Array[Byte](32)
     Random.nextBytes(forwardTransferHash)